+++ MobileGarage/android-screenshot-tests/src/screenshotTest/kotlin/com/chriscartland/garage/screenshottests/SettingsRedesignScreenshotTest.kt
@@ -32,6 +32,7 @@ import com.chriscartland.garage.ui.settings.SettingsContentSignedOutPreview
 import com.chriscartland.garage.ui.settings.SettingsContentSnoozeInFlightPreview
 import com.chriscartland.garage.ui.settings.SettingsContentWatchInstallPreview
 import com.chriscartland.garage.ui.settings.SettingsContentWatchInstalledPreview
+import com.chriscartland.garage.ui.settings.SettingsContentWatchInstalledUnknownVersionPreview
 import com.chriscartland.garage.ui.settings.SnoozeSheetContentActivePreview
 import com.chriscartland.garage.ui.settings.SnoozeSheetContentOffPreview
 import com.chriscartland.garage.ui.settings.VersionSheetContentPreview
@@ -131,6 +132,18 @@ fun SettingsContentWatchInstallPreviewTest() {
 @Composable
 fun SettingsContentWatchInstalledPreviewTest() {
     AppTheme { SettingsContentWatchInstalledPreview() }
+}
+
+@PreviewTest
+@Preview(showBackground = true, name = "Light")
+@Preview(
+    showBackground = true,
+    name = "Dark",
+    uiMode = android.content.res.Configuration.UI_MODE_NIGHT_YES,
+)
+@Composable
+fun SettingsContentWatchInstalledUnknownVersionPreviewTest() {
+    AppTheme { SettingsContentWatchInstalledUnknownVersionPreview() }
 }
 
 @PreviewTest